feat: switch to sqlite3 for storing directory tree
with the TreeNode()-style design in the previous version, we were
storing the entire directory tree in memory, which made using the fs
nearly impossible to use on smaller machines or with big libraries. Now,
the inode table is stored in the sqlite3 database along with all the tag
data and is reset on each mount.

FLAC/MP3 header information is now stored at the Operations class level
and operates as a mru cache for a single item. If any kind of threaded
access happens this will probably slow down things a lot

Logging is now a member of Operations. I don't know if this is better