feat: Add event listeners to update fs on db changes
Add event listeners for 'import' and 'item_removed' to update the
filesystem on changes. The 'database_change' would be more appropriate
to catch when metadata is changed and not just added or deleted, but it
is too eager on tasks like importing an album, where it will update the
inode db for each item in an album. A third partial update method could
improve this.

Move inode functions to module scope. The functions don't have anything
to do with the filesystem operations and we need to be able to call them
from the beetfs plugin context.

Revert to global logging, which seems to be a better pattern.